+++ tests/test_ui.rs
@@ -1,7 +1,7 @@
 // test_ui.rs
 
 // *************************************************************************
-// * Copyright (C) 2018-2019 Daniel Mueller (deso@posteo.net)              *
+// * Copyright (C) 2018-2020 Daniel Mueller (deso@posteo.net)              *
 // *                                                                       *
 // * This program is free software: you can redistribute it and/or modify  *
 // * it under the terms of the GNU General Public License as published by  *
@@ -45,29 +45,29 @@ use crate::common::unwrap_custom;
 
 #[test]
 fn correct_ids() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w1 = ui.add_widget(root, &mut |id, _cap| {
+  let w1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
   // And a container.
-  let c1 = ui.add_widget(root, &mut |id, _cap| {
+  let c1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
   // And a widget to the container.
-  let w3 = ui.add_widget(c1, &mut |id, _cap| {
+  let w3 = ui.add_ui_widget(c1, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
   // And another container for deeper nesting.
-  let c2 = ui.add_widget(c1, &mut |id, _cap| {
+  let c2 = ui.add_ui_widget(c1, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
   // And the last widget.
-  let w4 = ui.add_widget(c2, &mut |id, _cap| {
+  let w4 = ui.add_ui_widget(c2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -81,22 +81,22 @@ fn correct_ids() {
 
 #[test]
 fn creation_order_is_child_order() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w1 = ui.add_widget(root, &mut |id, _cap| {
+  let w1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w11 = ui.add_widget(w1, &mut |id, _cap| {
+  let w11 = ui.add_ui_widget(w1, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w12 = ui.add_widget(w1, &mut |id, _cap| {
+  let w12 = ui.add_ui_widget(w1, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w21 = ui.add_widget(w2, &mut |id, _cap| {
+  let w21 = ui.add_ui_widget(w2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -122,14 +122,14 @@ fn creation_order_is_child_order() {
 #[cfg(debug_assertions)]
 #[should_panic(expected = "The given Id belongs to a different Ui")]
 fn share_ids_between_ui_objects() {
-  let (mut ui1, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui1, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let widget = ui1.add_widget(root, &mut |id, _cap| {
+  let widget = ui1.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
-  let (mut ui2, _) = Ui::new(&mut |id, _cap| {
+  let (mut ui2, _) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -141,16 +141,16 @@ fn share_ids_between_ui_objects() {
 
 #[test]
 fn visibility_fun() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w1 = ui.add_widget(root, &mut |id, _cap| {
+  let w1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w3 = ui.add_widget(w2, &mut |id, _cap| {
+  let w3 = ui.add_ui_widget(w2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -206,14 +206,14 @@ fn visibility_fun() {
 
 #[test]
 fn no_initial_focus() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
   // No widget has the input focus by default.
   assert!(ui.focused().is_none());
 
-  let _ = ui.add_widget(root, &mut |id, _cap| {
+  let _ = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
   assert!(ui.focused().is_none());
@@ -221,10 +221,10 @@ fn no_initial_focus() {
 
 #[test]
 fn focus_widget() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let widget = ui.add_widget(root, &mut |id, _cap| {
+  let widget = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -234,10 +234,10 @@ fn focus_widget() {
 
 #[test]
 fn focus_makes_widget_visible() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let widget = ui.add_widget(root, &mut |id, _cap| {
+  let widget = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -255,25 +255,25 @@ fn focus_makes_widget_visible() {
 
 #[test]
 fn focus_changes_child_order() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w1 = ui.add_widget(root, &mut |id, _cap| {
+  let w1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w11 = ui.add_widget(w1, &mut |id, _cap| {
+  let w11 = ui.add_ui_widget(w1, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w21 = ui.add_widget(w2, &mut |id, _cap| {
+  let w21 = ui.add_ui_widget(w2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w22 = ui.add_widget(w2, &mut |id, _cap| {
+  let w22 = ui.add_ui_widget(w2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w23 = ui.add_widget(w2, &mut |id, _cap| {
+  let w23 = ui.add_ui_widget(w2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -314,19 +314,19 @@ fn focus_changes_child_order() {
 
 #[test]
 fn repeated_show_preserves_order() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w1 = ui.add_widget(root, &mut |id, _cap| {
+  let w1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w11 = ui.add_widget(w1, &mut |id, _cap| {
+  let w11 = ui.add_ui_widget(w1, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w21 = ui.add_widget(w2, &mut |id, _cap| {
+  let w21 = ui.add_ui_widget(w2, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -345,19 +345,19 @@ fn repeated_show_preserves_order() {
 
 #[test]
 fn hide_changes_child_order() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w1 = ui.add_widget(root, &mut |id, _cap| {
+  let w1 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w3 = ui.add_widget(root, &mut |id, _cap| {
+  let w3 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w4 = ui.add_widget(root, &mut |id, _cap| {
+  let w4 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -384,16 +384,16 @@ fn hide_changes_child_order() {
 
 #[test]
 fn repeated_hide_preserves_order() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let _ = ui.add_widget(root, &mut |id, _cap| {
+  let _ = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w2 = ui.add_widget(root, &mut |id, _cap| {
+  let w2 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let w3 = ui.add_widget(root, &mut |id, _cap| {
+  let w3 = ui.add_ui_widget(root, |id, _cap| {
     Box::new(TestWidget::new(id))
   });
 
@@ -429,20 +429,20 @@ struct CreatingWidget {
 
 impl CreatingWidget {
   pub fn new(id: Id, cap: &mut dyn MutCap<Event>) -> Self {
-    let child = cap.add_widget(id, &mut |id, _cap| {
+    let child = cap.add_widget(id, Box::new(|id, _cap| {
       let widget = TestWidgetBuilder::new()
         .custom_handler(counting_handler)
         .build(id);
       Box::new(widget)
-    });
+    }));
     // Focus the "last" widget. Doing so allows us to send an event to
     // all widgets.
     cap.focus(child);
 
     if need_more(id, cap) {
-      let _ = cap.add_widget(id, &mut |id, cap| {
+      let _ = cap.add_widget(id, Box::new(|id, cap| {
         Box::new(CreatingWidget::new(id, cap))
-      });
+      }));
     }
 
     CreatingWidget {
@@ -464,7 +464,7 @@ impl Handleable<Event> for CreatingWidget {
 fn recursive_widget_creation() {
   // We only create the root widget directly but it will take care of
   // recursively creating a bunch of more widgets.
-  let (mut ui, _) = Ui::new(&mut |id, cap| {
+  let (mut ui, _) = Ui::new(|id, cap| {
     Box::new(CreatingWidget::new(id, cap))
   });
 
@@ -501,13 +501,11 @@ impl MovingWidget {
 // ownership of some data during its construction.
 #[test]
 fn moving_widget_creation() {
-  let mut object = Some(Moveable {});
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     Box::new(TestWidget::new(id))
   });
-  let _ = ui.add_widget(root, &mut |id, _cap| {
-    let moveable = object.take().unwrap();
-    Box::new(MovingWidget::new(id, moveable))
+  let _ = ui.add_ui_widget(root, move |id, _cap| {
+    Box::new(MovingWidget::new(id, Moveable {}))
   });
 }
 
@@ -515,9 +513,9 @@ fn moving_widget_creation() {
 fn create_handler(widget: Id, event: Event, cap: &mut dyn MutCap<Event>) -> Option<UiEvents> {
   match event {
     Event::Key(key) if key == 'z' => {
-      cap.add_widget(widget, &mut |id, _cap| {
+      cap.add_widget(widget, Box::new(|id, _cap| {
         Box::new(TestWidget::new(id))
-      });
+      }));
       None
     },
     _ => Some(event.into()),
@@ -527,7 +525,7 @@ fn create_handler(widget: Id, event: Event, cap: &mut dyn MutCap<Event>) -> Opti
 
 #[test]
 fn event_based_widget_creation() {
-  let (mut ui, root) = Ui::new(&mut |id, _cap| {
+  let (mut ui, root) = Ui::new(|id, _cap| {
     let widget = TestWidgetBuilder::new()
       .event_handler(create_handler)
       .build(id);
@@ -559,7 +557,7 @@ fn recursive_operations_handler(widget: Id,
 
 #[test]
 fn recursive_widget_operations() {
-  let (mut ui, _) = Ui::new(&mut |id, _cap| {
+  let (mut ui, _) = Ui::new(|id, _cap| {
     let widget = TestWidgetBuilder::new()
       .custom_handler(recursive_operations_handler)
       .build(id);